+++ tests/testthat/test-fit_model_user_fit_function.R
@@ -2,7 +2,7 @@ library(testthat)
 suppressMessages(library(lavaan))
 library(manymome)
 
-skip_if_not_installed(lmhelprs)
+skip_if_not_installed("lmhelprs")
 
 test_that("fit_model: User fit function", {